requirements.txt
requests==2.21.0
pexpect==4.7.0
PyYAML==5.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
PyMySQL==0.9.3
setuptools==33.1.1
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
Jinja2==2.10.1